feat(sveltekit): Ensure AsyncLocalStorage AsyncContextStrategy is used in cloudflare pages
