+++ packages/cloudflare/src/async.ts
@@ -10,6 +10,9 @@ import { AsyncLocalStorage } from 'node:async_hooks';
  *
  * AsyncLocalStorage is only available in the cloudflare workers runtime if you set
  * compatibility_flags = ["nodejs_compat"] or compatibility_flags = ["nodejs_als"]
+ *
+ * @internal Only exported to be used in higher-level Sentry packages
+ * @hidden Only exported to be used in higher-level Sentry packages
  */
 export function setAsyncLocalStorageAsyncContextStrategy(): void {
   const asyncStorage = new AsyncLocalStorage<{
